changed types for simplicity of the example

+++ contracts/inherited-storage/LogicV2.sol
@@ -4,13 +4,13 @@ pragma solidity ^0.8.0;
 import "./LogicV1.sol";
 
 contract LogicV2 is LogicV1 {
-    uint8 internal age;
+    uint256 internal age;
 
-    function getAge() external view returns(uint8){
+    function getAge() external view returns(uint256){
         return age;
     }
 
-    function setAge(uint8 _age) external {
+    function setAge(uint256 _age) external {
         age = _age;
     }
 }
